move package to fix_base
