+++ src/phx/fix_base/fix/model/position_report.py
@@ -1,7 +1,7 @@
 from typing import List, Any
 from tabulate import tabulate
 
-from phx.fix.model.message import Message
+from phx.fix_base.fix.model.message import Message
 
 
 class Position:
