+++ src/phx/fix_base/fix/utils/fix.py
@@ -1,7 +1,7 @@
 import quickfix as fix
 
-from phx.utils.utils import str_to_datetime
-from phx.fix.model.auth import FixAuthenticationMethod
+from phx.fix_base.utils.utils import str_to_datetime
+from phx.fix_base.fix.model.auth import FixAuthenticationMethod
 
 
 def flip_trading_dir(direction):
